+++ source/deprecated/omni.isaac.debug_draw/omni/isaac/debug_draw/__init__.py
@@ -15,3 +15,5 @@ new_extension_name = "isaacsim.util.debug_draw"
 carb.log_warn(
     f"{old_extension_name} has been deprecated in favor of {new_extension_name}. Please update your code accordingly."
 )
+
+from isaacsim.utils.debug_draw import *