+++ source/deprecated/omni.isaac.manipulators/omni/isaac/manipulators/tests/test_imports.py
@@ -1,0 +1,24 @@
+# Import test file for omni.isaac.manipulators
+# This file was automatically generated to test imports from deprecated extensions
+# Compiled from tests found in python/tests
+
+import omni.kit.test
+
+
+class TestImports(omni.kit.test.AsyncTestCase):
+    async def setUp(self):
+        pass
+
+    async def tearDown(self):
+        pass
+
+    async def test_imports_for_omni_isaac_manipulators_extension(self):
+        # Testing all imports from original extension tests
+        import numpy as np
+        from omni.isaac.core import World
+        from omni.isaac.core.utils.stage import add_reference_to_stage, create_new_stage_async
+        from omni.isaac.manipulators import SingleManipulator
+        from omni.isaac.manipulators.grippers import SurfaceGripper
+        from omni.isaac.nucleus import get_assets_root_path_async
+
+        print("All imports successful for extension: omni.isaac.manipulators")