+++ source/deprecated/omni.isaac.utils/premake5.lua
@@ -4,6 +4,7 @@ project_ext(ext)
 
 repo_build.prebuild_link {
     { "python/scripts", ext.target_dir .. "/omni/isaac/utils/scripts" },
+    { "python/tests", ext.target_dir .. "/omni/isaac/utils/tests" },
     { "docs", ext.target_dir .. "/docs" },
     { "data", ext.target_dir .. "/data" },
 }