+++ source/deprecated/omni.isaac.sensor/docs/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [13.0.4] - 2025-03-25
+### Changed
+- Add import tests for deprecated extensions
 
 ## [13.0.3] - 2025-01-21
 ### Changed